refactor BE routes a little
