+++ src/hooks/index.ts
@@ -1,5 +1,4 @@
 export * from './useUser';
 export * from './useLists';
-export * from './useAllUsers';
 export * from './ThemeContext';
 export * from './useFriends';
